Add files via upload
This is the 2nd Task where we had to create a tic-tac-toe game. The logic used here is, a cell is chosen by the digits (0-8) for 9 cells (3X3 grid). User just needs to input a number and that corresponding cell would be filled. It uses minimax strategy and it's almost impossible to defeat the AI.